Explicitly separate core and plugin ABIs
