+++ include/core/utils/util.h
@@ -37,7 +37,7 @@ THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <stdbool.h>
 #include <stdint.h>   // Defines int64_t
 
-#include "plugin-apis/logging-abi.h"
+#include "plugin-apis/impl-in-core/logging-abi.h"
 
 /**
  * Holds generic statistical data
