+++ include/plugin-apis/impl-in-core/logging-abi.h
