+++ include/plugin-apis/impl-in-core/platform-abi.h
